+++ app/src/androidTest/java/com/minuk/pet/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.minuk.petcalendar
+package com.minuk.pet
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
