+++ app/src/main/java/com/minuk/pet/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.minuk.petcalendar
+package com.minuk.pet
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
