+++ app/src/test/java/com/minuk/pet/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.minuk.petcalendar
+package com.minuk.pet
 
 import org.junit.Test
